Migrated all content to add-business page
