+++ src/app/(product)/google-reviews/add-business/_components/step-indicator.tsx
